# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  domain.UserProfile:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      subcribers:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      subscribers_count:
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      subscriptions_count:
        type: integer
      updated_at:
        type: string
      username:
        type: string
      videos:
        items:
          $ref: '#/definitions/domain.Video'
        type: array
    type: object
  domain.Video:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      qualities:
        items:
          $ref: '#/definitions/domain.VideoQuality'
        type: array
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_url:
        type: string
      views:
        type: integer
    type: object
  domain.VideoQuality:
    properties:
      created_at:
        type: string
      id:
        type: integer
      quality:
        type: string
      video_id:
        type: integer
      video_url:
        type: string
    type: object
...
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  dto.SignInRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  dto.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.TokenPairResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  httpdelivery.CommonError:
    properties:
      error:
        type: string
    type: object
  httpdelivery.CommonResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: analogyoutube@example.com
    name: AnalogYouTube Backend
  description: |-
    API для AnalogYouTube: пользователи, видео, чаты, донаты и остальное.
    Для закрытых запросов нажмите Authorize и вставьте: Bearer <token>.
  title: AnalogYouTube API
  version: "1.0"
paths:
...
  /api/categories:
    get:
      description: Список категорий.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
      summary: Все категории
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Новая категория для видео.
      parameters:
      - description: Данные категории
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Создать категорию
      tags:
      - Categories
      - Admin
  /api/categories/{id}:
    delete:
      description: Удаляет категорию.
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Удалить категорию
      tags:
      - Categories
      - Admin
    put:
      consumes:
      - application/json
      description: Меняет название или описание категории.
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить категорию
      tags:
      - Categories
      - Admin
  /api/categories/{name}:
    get:
      description: Данные одной категории.
      parameters:
      - description: Имя категории
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Категория по имени
      tags:
      - Categories
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Возвращает access и refresh токены.
      parameters:
      - description: Данные входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Вход
      tags:
      - Auth
  /api/me:
    get:
      description: Данные текущего пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Мой профиль
      tags:
      - Profile
    put:
      consumes:
      - multipart/form-data
      description: Меняет мои данные, аватарку или пароль.
      parameters:
      - description: Новое имя пользователя
        in: formData
        name: username
        required: true
        type: string
      - description: Email, например user@gmail.com
        in: formData
        name: email
        type: string
      - description: Новый пароль
        in: formData
        name: password
        type: string
      - description: Описание канала
        in: formData
        name: description
        type: string
      - description: Файл аватарки
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить профиль
      tags:
      - Profile
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя. Email можно не указывать.
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Регистрация
      tags:
      - Auth
...
  /auth/refresh:
    get:
      description: Меняет refresh token на новую пару токенов.
      parameters:
      - description: Refresh token
        in: header
        name: Refresh-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenPairResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Обновить токены
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Возвращает access и refresh токены.
      parameters:
      - description: Данные входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Вход
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя. Email можно не указывать.
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Регистрация
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"